requirements.txt
tensorflow
discord.py
requests
loguru
absl-py
